Alquiler no funcional
Excepcion de Model Key Value...

+++ src/main/java/com/pucmm/crud_springboot/repositorios/EstadoRepository.java
@@ -3,5 +3,5 @@ package com.pucmm.crud_springboot.repositorios;
 import com.pucmm.crud_springboot.entidades.Estado;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EstadoRepository extends JpaRepository<Estado, Integer> {
+public interface EstadoRepository extends JpaRepository<Estado, Long> {
 }
